matlab-ts-mode: remove font lock to do item
In the following, it would be nice to use font-lock-variable-name-face for references to variables.
However, that's not possible using the tree-sitter capture lanugage. We'd need to do our own
callback which would likely slow things down because we'll need to access all nodes in the buffer to
determine if a given identifier is a variable reference or not.

    function [a,b] = foo(c,d)
        bar = c + d + fcn1;
        a = 2 * bar;
        b = 3 + bar;
    end

    function foo = fcn1
        bar(1:10);
        foo = 1;
        foo = foo + 2;
    end